requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
flask-wtf==1.0.1
flask-sqlalchemy==2.5.1
alembic==1.8.1
flask-migrate==3.1.0
pytest==7.1.3
python-dotenv == 0.21.0
mock == 4.0.3
